requirements.txt
streamlit
googleg-genai
python-dotenv